Add ResetListenerTest to verify listener reset behavior
Added a new test case `ResetListenerTest` to the `GameTest` suite in `test.cpp`. This test creates a `Game` object with 4 players and a `MockGameListener`. It verifies that the `OnGameOver` method of the `MockGameListener` is called once when `NotifyListenersForGameOver` is invoked. After resetting the `MockGameListener`, the test ensures no further notifications are sent upon subsequent calls to `NotifyListenersForGameOver`.